minor refactor: rename `code-block-controls` => `code-blocks`
