+++ src/Layout.tsx
@@ -15,7 +15,7 @@ import { ExternalLinks } from './ExternalLinks'
 import { coseMenuModalOnMouseLeave, openMenuModal, toggleMenuModal } from './MenuModal/toggleMenuModal'
 import { MenuModal } from './MenuModal'
 import { autoScrollNav_SSR } from './autoScrollNav'
-import { initializeJsToggle_SSR } from './code-block-controls/hooks/useSelectCodeLang'
+import { initializeJsToggle_SSR } from './code-blocks/hooks/useSelectCodeLang'
 import { SearchLink } from './docsearch/SearchLink'
 import { navigate } from 'vike/client/router'
 import { css } from './utils/css'
@@ -296,12 +296,12 @@ function NavHead({ isNavLeft }: { isNavLeft?: true }) {
         ...(!isNavLeft
           ? {}
           : {
-              position: 'absolute',
-              left: '100%',
-              top: 0,
-              '--padding-side': '20px',
-              width: mainViewMax, // guaranteed real estate
-            }),
+            position: 'absolute',
+            left: '100%',
+            top: 0,
+            '--padding-side': '20px',
+            width: mainViewMax, // guaranteed real estate
+          }),
       }}
     >
       {pageContext.globalContext.config.docpress.topNavigation}
